[components] Dynamic components receive capabilities from Dict
Components can be now created and started with a Dict that is used as
the source for `use`d capabilities.

This allows a component to dynamically route capabilities for every
execution (via CreateChildArgs.dict), and for a specific execution
(via StartChildArgs.dict).

Compared to a dynamic offers:

  * The capability source can vary per-component, especially for
    protocols. With dynamic offers, this is only possible with
    directory/service capabilities. Dicts can provide a unique
    implementation for protocols.
  * The capability source can be set for a single component
    execution. Dynamic offers persist across component restarts,
    forcing the parent to recreate the component.

When a component connects to a capability and the Dict contains an
Open value with a key that matches the capability name, it will be
provided from the Dict.

Routing from a dict is incompatible with static routes. If a capability
in a Dict is already routed statically, the Dict will be rejected.

Bug: b/301973404
Change-Id: I48ba37b0930d65881ccb0893d26db8b4335a666c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931533
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Gary Bressler <geb@google.com>