[starnix][iptables] Parse ipt_replace entries
This CL peeks inside the entries field and attempts to create structures
that the fuchsia.net.filter API expects. It does not call the API to
commit these structures into the system, but it validates that the bytes
are set properly, and returning EINVAL otherwise.

IPTable rules were and continue to take no effect on the system.

Bug: b/307908515
Change-Id: Ic37433a6aa43302764f795017f440d28d79439c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946613
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: David Song <wintermelons@google.com>