[driver][sdk] Improvements to compat DeviceServer
This improves the ergonomics of the DeviceServer's self initialization
logic a bit. Previously it was confusing whether the async or sync init
was happening. And the callback mechanism was also confusing. This
creates a wrapper type for each sync/async inialization so it is clear
exactly what is happening.

Bug: 314040286

Change-Id: I1453b5d7d2be06d8e8047540ead29bb8dfe44cb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979517
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>