[superproject] Roll third_party/pigweed pw_ide: Remove terminal env vars from VSC settings
As of pwrev.dev/187195, an activated terminal can be created through
a VSC command, which ensures that the activation env vars are always
up-to-date. So we no longer need to populate them in `settings.json`
(and risk them going stale).

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/188470
Original-Revision: 76c3f705fb7e8f295140136bfd33d1a2d32977aa
GitOrigin-RevId: 3b71dfc4ffa5f2e0fa67dbfb71c592f48e219af2
Roller-URL: https://ci.chromium.org/b/8757528500410010721
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id80da1c76c008a041f9414034ad8664e088082d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982593