[kernel][gicv2] add a spinlock to protect irq configuration
In the case of gic_configure and gic_set_affinity a RMW sequence needs
to be lock in the extraordinarily rare case where two cpus tries to
fiddle with the same line. Add a spinlock to protect.

Also replaced static with anonymous namespace, which is the bulk of the
file's changes.
Code formatter picked up a few stray old changes that hadn't been
touched in a while.

Bug: b/320570343
Change-Id: I2f2886127c67068849be52cc0d3366885593cbe0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975100
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>