[ffx][flash]: flash uses event based reconnect
This is a reland of commit 10423420caf91294be7e09c6134ac576376962b3

The UsbFactory's rediscover method needed to be refactored from the
original commit to not cause crosstalk.

Using the generic "wait_for_devices" function, each usb
device is probed to see if it is in fastboot mode. While this works fine
in low volume situations. In an infrastructure environment with more
than 10 devices idling in fastboot (or being flashed actively) at a
time, this can cause crosstalk and unexpected errors. By doing a
more strict discovery of USB Targets, we can eliminate this.

This adds function check to usb_bulk::AsyncInterface, which returns true
if an Interface could be opened which matches the provided matcher.
it should be noted that this does NOT drain the buffer which
usb_bulk::AsyncInterface::open will do.

Now, in the flash process, when rebooting the bootloader (after flashing
the new bootloader) instead of using a configurable sleep to wait for
the target to reappear (and continue to flash the product), the tool
will use the discovery library to begin looking for the target to
reappear in fastboot mode.

This assumes:
  1. A target will never change fastboot connection mechanisms. For
     example, it is assumed that if a target is connected via USB, after
     the bootloader reboots it will continue to be available via USB
  2. A target's name (for fastboot over TCP/UDP) is consistent across
     reboots.
  3. A target's serial number (for fastboot over USB) is consistent
     across reboots.

Release-Notes: ffx uses event based reconnect for flashing.
Change-Id: I9296fb8b9d15002faca2b64d653432eb1cf47ed7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974920
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>