[unwinder] Update frame count check
Instead of estimating the number of total frames that should be present
after unwinding, we now find the frame corresponding to the start of
the test and verify that the required minimum number of frames
corresponding to the nested calls made are present. This avoids issues
that occur with varying frame counts due to compile time and link
time optimizations.

Fixed: 318622080
Tests: fx test unwinder_tests
Change-Id: Ibcebeacb0c94e5b90e357f3c9079a564bac73448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983967
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>