[wlansoftmac] Add async Tx path tracing
In general, when tracing is off, each tracing call in the code is
"about the cost of a branch instruction" which is on the order of
nanoseconds. When tracing is on, it should cost a hashtable lookup
plus some small memcpy's which is on the order of <200 nanoseconds. In
other words, with tracing on or off, we would not expect tracing to
bottleneck data path performance until we're transmitting at very high
data rates.

Test: Manually recorded a trace
Change-Id: Ic2ca039402cd78ef7e90a1372bb8d4bbadb1ecdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965580
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>