[component_manager] Update LaunchTask to use Router
Update LaunchTask to use Router that returns an Open.
This removes the last amount of generics from
Sender and Receiver.

The main logical change of this CL is that LaunchTask
no longer needs to be served from an async TaskGroup.
This is because it is now stored in the Dictionary and
called directly. This simplifies some of the code.

Bug: b/310741884
Change-Id: I130e3d29f82528c7a53af8c2d487eed058b4e7b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983356
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>