[debugger] Mode SymbolFactory to the DwarfBinary.
Previously the DwarfBinary (representing the file on disk) and the
SymbolFactory (responsible for decoding stuff in the binary) were
separate. They were stitched together in the ModuleSymbolsImpl
class.

This made it impossible to decode symbols given a DwarfBinary which
is all some classes have like DwarfUnit (this is at a lower layer than
ModuleSymbols). This capability is needed in a followup for Debug
Fission support.

This patch moves ownership of the symbol factory to the DwarfBinary.

Change-Id: Ifbf51d128c6d7e7cf04870610cc79bcce9a61e7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977131
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>