[ffx][usb_fastboot_discovery]: loosen device product requirements
Previously, the check of the USB Interface to detect if it was a valid
Device in Fastboot was very strict, and would require matching on Vendor
and Product for the Interface.

Now, loosen the restriction for device products to be only a warning.

Change-Id: I0f80afbd7ad1382c81cddce00f752b3471533450
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985098
Reviewed-by: Jiaming Li <lijiaming@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>