[ffx][usb_fastboot_discovery]: add diagnostics
Adds diagnostics context to open_interface_with_serial

Add struct UnversionedFastbootUsbTester, which will check if the Usb
Device with the provided serial number is fastboot over useb. This
differs from the OpenInterfaceFastbootUsbTester in that it does not send
the fastboot GetVar command to it to verify that the interface is up and
"Live" only that the USB Device is present.

Change-Id: I02c1a552385a6d896dd92d235d0b731b06ece857
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984400
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>