[fuchsia-repo] Add support for subpackages
This change updates fuchsia-repo client to support recursively
enumerating the subpackages of a package. This is needed to fix a
problem with pbtool, which was not including the subpackage blobs in the
generated product bundle transfer manifests. This also allows
list_packages() to take in account the subpackages when computing the
total package size.

The `ffx repository package show` command was also updated to optionally
display subpackage information using the new --include-subpackages flag.

Bug: b/319394340
Change-Id: I894ffb1d490518a33eb0a1a25602e9d3b1fe6d15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979877
Commit-Queue: Michael Brase <mbrase@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>